Options runtime checks fix.
`options` parameter is defined as optional in both `getSourceJsonUrl`
and `getSourcesData`.
However in implementation it is treated as if it guaranteed to be
present and compile checks are overriden by `!!`.
As a result, [example code](https://github.com/ton-community/contract-verifier-sdk/blob/main/example/node/index.ts)
is guaranteed to fail at runtime.